feat(router): add pure routing subpath
